fix: 퀴즈 한문제 종류후 퀴즈 시작 눌렀을때 학생 페이지 오류 해결
